implemented authorization

+++ app/controllers/products_controller.ts
@@ -3,16 +3,21 @@ import { cuid } from '@adonisjs/core/helpers'
 
 import Product from '#models/product'
 import app from '@adonisjs/core/services/app'
-import Image from '#models/image'
+import { isAdmin } from '#abilities/main'
+import Image from '#models/image' 
 
 import { createProductValidator } from '#validators/product'
 
 export default class ProductsController {
-  public async index({ view }: HttpContext) {
+  public async index({ view, bouncer, response }: HttpContext) {
+    if(!(await bouncer.allows(isAdmin))) {
+      return response.status(403).send('Not authorized')
+    }
     const products = await Product.all()
     for (const product of products) {
       await product.load('images')
     }
+    
     return view.render('pages/products/index', { products })
   }
 
@@ -22,17 +27,28 @@ export default class ProductsController {
     return view.render('pages/products/show', { product, image })
   }
 
-  public async create({ view }: HttpContext) {
+  public async create({ view, bouncer, response}: HttpContext) {
+    if(!(await bouncer.allows(isAdmin))) {
+      return response.status(403).send('Not authorized')
+    }
     return view.render('pages/products/create')
   }
 
-  public async edit({ params, view }: HttpContext) {
+  public async edit({ params, view, bouncer, response }: HttpContext) {
     const product = await Product.findOrFail(params.id)
+
+    if(!(await bouncer.allows(isAdmin))) {
+      return response.status(403).send('Not authorized')
+    }
 
     return view.render('pages/products/create', { product })
   }
 
-  public async store({ request, response }: HttpContext) {
+  public async store({ request, response, bouncer }: HttpContext) {
+    if(!(await bouncer.allows(isAdmin))) {
+      return response.status(403).send('Not authorized')
+    }
+
     const payload = await request.validateUsing(createProductValidator)
     console.log(payload, "validated")
     const product = await Product.create({
@@ -54,7 +70,10 @@ export default class ProductsController {
     return response.redirect().toRoute('products.show', { id: product.id })
   }
 
-  public async update({ params, request, response }: HttpContext) {
+  public async update({ params, request, response, bouncer }: HttpContext) {
+    if(!(await bouncer.allows(isAdmin))) {
+      return response.status(403).send('Not authorized')
+    }
     const product = await Product.findOrFail(params.id)
 
     const payload = await request.validateUsing(createProductValidator)
@@ -65,7 +84,10 @@ export default class ProductsController {
     return response.redirect().toRoute('products.show', { id: product.id })
   }
 
-  public async destroy({ params, response }: HttpContext) {
+  public async destroy({ params, response, bouncer }: HttpContext) {
+    if(!(await bouncer.allows(isAdmin))) {
+      return response.status(403).send('Not authorized')
+    }
     const product = await Product.findOrFail(params.id)
 
     await product.delete()
